% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_exact_2x2.R
\name{meta_exact_2x2}
\alias{meta_exact_2x2}
\title{Exact inference for meta-analysis of 2x2 tables}
\usage{
meta_exact_2x2(
  dat,
  root.tol = 1e-12,
  lambdas = NULL,
  lb.interval = NULL,
  ub.interval = NULL
)
}
\arguments{
\item{dat}{A data.frame with one row per study and columns as specified in the Details.}

\item{root.tol}{Tolerance for checking that the imaginary components of the polynomial roots are not too far from 0; default \code{= 1e-12}. See Details.}

\item{lambdas}{Optional input to provide the lambdas (polynomial roots) for some studies. Should be a list object where the kth element is a vector of lambdas for the kth study in dat and NULL for other studies. The default (\code{NULL}) computes the roots for all studies. See Details.}

\item{lb.interval}{Optional input to set the interval for \code{uniroot} to solve the Blaker lower bound; must be a length 2 vector if used. The default (\code{NULL}) is to guess a reasonable interval based on the Fisher Information based lower bound.}

\item{ub.interval}{Optional input to set the interval for \code{uniroot} to solve the Blaker upper bound; must be a length 2 vector if used. The default (\code{NULL}) is to guess a reasonable interval based on the Fisher Information based upper bound.}
}
\value{
Returns a named matrix with 2 rows and 4 columns. The first row contains the values for the exact inference using the Blaker method, and the second row contains the values using the Fisher Information approach. The columns are:
\item{est}{The cMLE log odds ratio estimate}
\item{ci.lb}{The lower bound of the 95\% confidence interval}
\item{ci.ub}{The upper bound of the 95\% confidence interval}
\item{p}{The p-value}
}
\description{
A function to compute exact confidence intervals and p-values for
meta-analysis of 2x2 tables. The function solves for the cMLE of the log odds
ratio and returns exact confidence intervals and p-values using the Blaker
method as well as confidence intervals and p-values using the Fisher
Information.
}
\details{
The \code{dat} data.frame must have the following 4 columns with the
specified names:
\itemize{
\item xk: count in the treatment group with the outcome
\item tk: total count with the coutcome across both groups (treatment and
control)
\item mk: count in the treatment group
\item nk: count in the control group
}

The function computes polynomial roots of a hypergeometric distribution for
each study. In some instances (e.g. for large studies), the R \code{polyroot}
function may have difficulty solving these roots, as indicated by non-zero
imaginary components (the \code{root.tol} parameter sets the tolerance for
identifying non-zero components.) If this occurs, you can either remove the
problematic study from the meta-analysis or calculate the polynomial roots
using another software (e.g. Mathematica) and import them using the
'\code{lambdas} input to this function.

This optional \code{lambdas} input allows for importing roots for a subset of
studies; it should be a list object with length equal to the number of rows
(studies) in \code{dat}, where the kth element is a vector of lambdas for the
kth study in \code{dat}. For every element of \code{lambdas} where a vector
of roots is provided, the function will use those roots; for every element of
\code{lambdas} that is \code{NULL}, the function will compute the polynomial
roots using the \code{polyroot} function.
}
